+++ ExamplePlugin/ThunderstoreMeta/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
 
+## 0.1.2
+
+- Resolve non-host clients not sending message to the server
+  - On stage start, the server requests a LoadoutSync from the clients
+
+## 0.1.1
+
+- fixing links in the readme by using imgur source
+
 ## 0.1.0
 
 - Initial Release